Rename db to database
